Rewrote Server Condition in TS
